fix: inoperable troop adjustments removed from farm list
The interface, along with the in-line JavaScript code on the page was
most probably taken from the actual 3.5 real game and never tested.
Therefore, we're missing a whole JavaScript file of functionality that
would update units in raid lists.

This would be too much work to do at this stage, so I'm shelving this
functionality.